refactor: extract camera-ready config helpers
